+++ CHANGELOG.md
@@ -7,6 +7,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+- Add metrics from `/proc/[pid]/io`
+
 ### Changed
 
 ### Deprecated